Add code for grouping and analyzing car sales data
Added a code that groups the car sales data by colour and plots it. This allows us to analyze the car sales data based on colour which might give us some interesting insights. Also, the head of the data frame is displayed for quick checking. The motivation behind this change is to provide a way to discover potential trends and correlations in the data related to car colours.