"Add Pandas data manipulation exercises
A new section titled "Manipulating data with Pandas" has been added to the "intro-to-pandas" notebook. This includes explanations and examples on how to manipulate data with Pandas library, methods like head(), fillna() and others were used. Also, a dataset with missing values was imported and used for the demonstration of handling such cases."