Add new features to dataframe
A new section was added to the pandas tutorial that demonstrates how to add new features (columns) to an existing pandas dataframe. A new column "Seats" was created by converting series to columns and then NaN values were filled with five. Second, a new column "Fuel per 100KM" was created by converting a list to a column. Finally, a second "Fuel per 100KM" column was created from a pandas series, which results in NaN values for unmatched indices.